Updated Building Comparison Pages. Updated individual building page to pass current building name to the picker
